Benwaa

+++ client.lua
@@ -5,9 +5,11 @@ function SaveX(sErr)
     end
     file.remove("s.txt")
     file.open("s.txt","w+")
+    print("Saving new config:")
     for k, v in pairs(s) do
+        print(k .. "=" .. v)
         file.writeline(k .. "=" .. v)
-    end                
+    end
     file.close()
     collectgarbage()
 end
@@ -18,8 +20,11 @@ function mysplit(inputstr, sep)
         end
         local t={} ; i=1
         for str in string.gmatch(inputstr, "([^"..sep.."]+)") do
-                t[i] = str
-                i = i + 1
+                -- Forces the files to contain a dot in the name.
+                if string.match(str, "%.") then
+                    t[i] = str
+                    i = i + 1
+                end
         end
         return t
 end
@@ -28,14 +33,13 @@ function dwn()
     -- body
     n = n + 1
     v = data[n]
-    if v == nil then 
+    if v == nil then
         --dofile(data[1]..".lc")
         bootfile= string.gsub(data[1], '\.lua$','') --string.gsub(s, '\....$','')
         s.boot = bootfile..".lc"
         SaveX("No error")
-        node.restart()
-
-    else 
+        Reboot("File loaded Reboot...")
+    else
         print("Filename: "..v)
         filename=v
 
@@ -43,7 +47,7 @@ function dwn()
             file.open(v, "w+")
 
             payloadFound = false
-            conn=net.createConnection(net.TCP, false) 
+            conn=net.createConnection(net.TCP, 0)
             conn:on("receive", function(conn, payload)
 
                 if (payloadFound == true) then
@@ -60,7 +64,7 @@ function dwn()
                 payload = nil
                 collectgarbage()
             end)
-            conn:on("disconnection", function(conn) 
+            conn:on("disconnection", function(conn)
                 conn = nil
                 file.close()
                 ext = string.sub(v, -3)
@@ -71,17 +75,24 @@ function dwn()
 
             end)
             conn:on("connection", function(conn)
-                conn:send("GET /"..s.path.."/uploads/"..id.."/"..v.." HTTP/1.0\r\n"..
+                cmd = "GET /"..s.path.."/uploads/"..id.."/"..v.." HTTP/1.0\r\n"..
                       "Host: "..s.host.."\r\n"..
                       "Connection: close\r\n"..
                       "Accept-Charset: utf-8\r\n"..
                       "Accept-Encoding: \r\n"..
-                      "User-Agent: Mozilla/4.0 (compatible; esp8266 Lua; Windows NT 5.1)\r\n".. 
-                      "Accept: */*\r\n\r\n")
+                      "User-Agent: Mozilla/4.0 (compatible; esp8266 Lua; Windows NT 5.1)\r\n"..
+                      "Accept: */*\r\n\r\n"
+                print("Request_HTTP:"..cmd)
+                conn:send(cmd)
             end)
             conn:connect(80,s.host)
     end
 
+end
+
+function Reboot(msg)
+    print("Restarting: "..msg)
+    node.restart()
 end
 
 function FileList(sck,c)
@@ -94,18 +105,19 @@ function FileList(sck,c)
     print("length: "..string.len(c))
 
     data = mysplit(c, "\n") -- fill the field with filenames
-    
+
     --for k,v in pairs(data) do
     n = 1
     v = data[n]
+    if v == nil then Reboot("No file to download") end
         print("Filename: "..v)
         filename=v
-        
+
             file.remove(v);
             file.open(v, "w+")
 
             payloadFound = false
-            conn=net.createConnection(net.TCP, false) 
+            conn=net.createConnection(net.TCP, 0)
             conn:on("receive", function(conn, payload)
 
                 if (payloadFound == true) then
@@ -122,7 +134,7 @@ function FileList(sck,c)
                 payload = nil
                 collectgarbage()
             end)
-            conn:on("disconnection", function(conn) 
+            conn:on("disconnection", function(conn)
                 conn = nil
                 file.close()
                 ext = string.sub(v, -3)
@@ -132,18 +144,17 @@ function FileList(sck,c)
                 dwn()
             end)
             conn:on("connection", function(conn)
-                conn:send("GET /"..s.path.."/uploads/"..id.."/"..v.." HTTP/1.0\r\n"..
+                cmd = "GET /"..s.path.."/uploads/"..id.."/"..v.." HTTP/1.0\r\n"..
                       "Host: "..s.host.."\r\n"..
                       "Connection: close\r\n"..
                       "Accept-Charset: utf-8\r\n"..
                       "Accept-Encoding: \r\n"..
-                      "User-Agent: Mozilla/4.0 (compatible; esp8266 Lua; Windows NT 5.1)\r\n".. 
-                      "Accept: */*\r\n\r\n")
+                      "User-Agent: Mozilla/4.0 (compatible; esp8266 Lua; Windows NT 5.1)\r\n"..
+                      "Accept: */*\r\n\r\n"
+                print("Request_HTTP:"..cmd)
+                conn:send(cmd)
             end)
             conn:connect(80,s.host)
-
-
-
 
     --end
     collectgarbage()
@@ -156,7 +167,15 @@ filename=nil
 LoadX()
 
 wifi.setmode (wifi.STATION)
-wifi.sta.config(s.ssid, s.pwd)
+station_cfg={}
+if  s.ssid ~= nil or s.ssid ~= '' then
+    station_cfg.ssid=s.ssid
+    station_cfg.save=true
+end
+if  s.pwd ~= nil or s.pwd ~= '' then
+    station_cfg.pwd=s.pwd
+end
+wifi.sta.config(station_cfg)
 wifi.sta.autoconnect (1)
 
 iFail = 20 -- trying to connect to AP in 20sec, if not then reboot
@@ -164,10 +183,11 @@ tmr.alarm (1, 1000, 1, function ( )
   iFail = iFail -1
   print(iFail)
   if (iFail == 0) then
-    SaveX("could not access "..s.ssid)
-    node.restart()
-  end      
-  
+    msg = "could not access "..s.ssid
+    SaveX(msg)
+    Reboot(msg)
+  end
+
 
   if wifi.sta.getip ( ) == nil then
     print(s.ssid..": "..iFail)
@@ -176,24 +196,26 @@ tmr.alarm (1, 1000, 1, function ( )
     tmr.stop (1)
     -- get list of files
     sk=net.createConnection(net.TCP, 0)
+    cmd = "GET /".. s.path .."/node.php?id="..id.."&list"..
+        " HTTP/1.1\r\n"..
+        "Host: "..s.domain.."\r\n"..
+        "Accept: */*\r\n"..
+        "User-Agent: Mozilla/4.0 (compatible; esp8266 Lua;)"..
+        "\r\n\r\n"
+    print("Request_HTTP:"..cmd)
     sk:on("connection",function(conn, payload)
-                sk:send("GET /".. s.path .."/node.php?id="..id.."&list"..
-                " HTTP/1.1\r\n".. 
-                "Host: "..s.domain.."\r\n"..
-                "Accept: */*\r\n"..
-                "User-Agent: Mozilla/4.0 (compatible; esp8266 Lua;)"..
-                "\r\n\r\n") 
+                sk:send(cmd)
             end)
     sk:on("receive", FileList)
-    
+
     --sGet = "GET /".. s.path .. " HTTP/1.1\r\nHost: " .. s.domain .. "\r\nConnection: keep-alive\r\nAccept: */*\r\n\r\n"
-    sk:connect(80,s.host) 
-    
+    sk:connect(80,s.host)
+
   end
   collectgarbage()
- 
+
 end)
 
 
 
- print(collectgarbage("count").." kB used")
+print(collectgarbage("count").." kB used")
